// src/lex-api/types/chat/bsky/convo/defs.ts
const id = 'chat.bsky.convo.defs'

export interface MessageRef {
  $type?: 'chat.bsky.convo.defs#messageRef'
  did: string
  convoId: string
  messageId: string
}

const hashMessageRef = 'messageRef'

export function isMessageRef<V>(v: V) {
  return is$typed(v, id, hashMessageRef)
}

export function validateMessageRef<V>(v: V) {
  return validate<MessageRef & V>(v, id, hashMessageRef)
}

export interface MessageInput {
  $type?: 'chat.bsky.convo.defs#messageInput'
  text: string
  /** Annotations of text (mentions, URLs, hashtags, etc) */
  facets?: AppBskyRichtextFacet.Main[]
  embed?: $Typed<AppBskyEmbedRecord.Main> | { $type: string }
}

const hashMessageInput = 'messageInput'

export function isMessageInput<V>(v: V) {
  return is$typed(v, id, hashMessageInput)
}

export function validateMessageInput<V>(v: V) {
  return validate<MessageInput & V>(v, id, hashMessageInput)
}

export interface MessageView {
  $type?: 'chat.bsky.convo.defs#messageView'
  id: string
  rev: string
  text: string
  /** Annotations of text (mentions, URLs, hashtags, etc) */
  facets?: AppBskyRichtextFacet.Main[]
  embed?: $Typed<AppBskyEmbedRecord.View> | { $type: string }
  /** Reactions to this message, in ascending order of creation time. */
  reactions?: ReactionView[]
  sender: MessageViewSender
  sentAt: string
}

const hashMessageView = 'messageView'

export function isMessageView<V>(v: V) {
  return is$typed(v, id, hashMessageView)
}

export function validateMessageView<V>(v: V) {
  return validate<MessageView & V>(v, id, hashMessageView)
}

export interface DeletedMessageView {
  $type?: 'chat.bsky.convo.defs#deletedMessageView'
  id: string
  rev: string
  sender: MessageViewSender
  sentAt: string
}

const hashDeletedMessageView = 'deletedMessageView'

export function isDeletedMessageView<V>(v: V) {
  return is$typed(v, id, hashDeletedMessageView)
}

export function validateDeletedMessageView<V>(v: V) {
  return validate<DeletedMessageView & V>(v, id, hashDeletedMessageView)
}

export interface MessageViewSender {
  $type?: 'chat.bsky.convo.defs#messageViewSender'
  did: string
}

const hashMessageViewSender = 'messageViewSender'

export function isMessageViewSender<V>(v: V) {
  return is$typed(v, id, hashMessageViewSender)
}

export function validateMessageViewSender<V>(v: V) {
  return validate<MessageViewSender & V>(v, id, hashMessageViewSender)
}

export interface ReactionView {
  $type?: 'chat.bsky.convo.defs#reactionView'
  value: string
  sender: ReactionViewSender
}

const hashReactionView = 'reactionView'

export function isReactionView<V>(v: V) {
  return is$typed(v, id, hashReactionView)
}

export function validateReactionView<V>(v: V) {
  return validate<ReactionView & V>(v, id, hashReactionView)
}

export interface ReactionViewSender {
  $type?: 'chat.bsky.convo.defs#reactionViewSender'
  did: string
}

const hashReactionViewSender = 'reactionViewSender'

export function isReactionViewSender<V>(v: V) {
  return is$typed(v, id, hashReactionViewSender)
}

export function validateReactionViewSender<V>(v: V) {
  return validate<ReactionViewSender & V>(v, id, hashReactionViewSender)
}

export interface MessageAndReactionView {
  $type?: 'chat.bsky.convo.defs#messageAndReactionView'
  message: MessageView
  reaction: ReactionView
}

const hashMessageAndReactionView = 'messageAndReactionView'

export function isMessageAndReactionView<V>(v: V) {
  return is$typed(v, id, hashMessageAndReactionView)
}

export function validateMessageAndReactionView<V>(v: V) {
  return validate<MessageAndReactionView & V>(v, id, hashMessageAndReactionView)
}

export interface ConvoView {
  $type?: 'chat.bsky.convo.defs#convoView'
  id: string
  rev: string
  members: ChatBskyActorDefs.ProfileViewBasic[]
  lastMessage?:
    | $Typed<MessageView>
    | $Typed<DeletedMessageView>
    | $Typed<MessageAndReactionView>
    | { $type: string }
  muted: boolean
  status?: 'request' | 'accepted' | (string & {})
  unreadCount: number
}